add two Debian nodes
At this point, we're creating two nodes that we still can't access.
We're just defining what's coming next.
Those are two containers running a Python 3 distribution because Ansible needs Python 3 to be installed.
We're running a command to keep the containers alive after started.
And we're exposing some ports so we can later install Nginx.