create ansible config file
We need two files to start using Ansible.
ansible.cfg defines an "inventory", which lists the machines we want to control.
We also set the secrets parameter.
We then create a "hosts" files to list our nodes.
Since we're using docker, we need to use the docker container names as hosts, and add the ansible_connection parameter.